refactor: change archtecture and implement new interfaces
