+++ src/main/java/com/rflpazini/rogue/app/dataprovider/impl/CustomerRepository.java
@@ -1,8 +1,8 @@
 package com.rflpazini.rogue.app.dataprovider.impl;
 
-import com.rflpazini.rogue.app.dataprovider.CrudPattern;
+import com.rflpazini.rogue.domain.dataprovider.CrudPattern;
 import com.rflpazini.rogue.app.dataprovider.DataBase;
-import com.rflpazini.rogue.app.dataprovider.model.Customer;
+import com.rflpazini.rogue.app.entrypoint.model.Customer;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Hashtable;
