+++ src/main/java/com/rflpazini/rogue/app/entrypoint/model/Money.java
@@ -1,4 +1,4 @@
-package com.rflpazini.rogue.app.dataprovider.model;
+package com.rflpazini.rogue.app.entrypoint.model;
 
 import lombok.Data;
