+++ src/main/java/com/rflpazini/rogue/domain/dataprovider/QueueProvider.java
@@ -1,0 +1,5 @@
+package com.rflpazini.rogue.domain.dataprovider;
+
+public interface QueueProvider<T> {
+  void post(T entity);
+}
