+++ src/main/java/com/rflpazini/rogue/domain/usecase/AccountService.java
@@ -1,17 +1,17 @@
 package com.rflpazini.rogue.domain.usecase;
 
-import com.rflpazini.rogue.app.dataprovider.impl.CustomerRepository;
-import com.rflpazini.rogue.app.dataprovider.model.Account;
-import com.rflpazini.rogue.app.dataprovider.model.Customer;
+import com.rflpazini.rogue.app.entrypoint.model.Account;
+import com.rflpazini.rogue.app.entrypoint.model.Customer;
+import com.rflpazini.rogue.domain.dataprovider.CrudPattern;
 import javax.inject.Inject;
 import javax.ws.rs.BadRequestException;
 
 public class AccountService {
 
-  @Inject CustomerRepository customerRepository;
+  @Inject CrudPattern<Customer> repository;
 
   public Account getAccountInfo(String id) {
-    Customer customerAccount = customerRepository.findById(id);
+    Customer customerAccount = repository.findById(id);
 
     return Account.builder()
         .balance(customerAccount.getBalance())
@@ -20,24 +20,24 @@ public class AccountService {
   }
 
   public Account deposit(String id, double amount) {
-    Customer customerAccount = customerRepository.findById(id);
+    Customer customerAccount = repository.findById(id);
     double newBalance = customerAccount.getBalance() + amount;
 
     customerAccount.setBalance(newBalance);
-    customerRepository.update(customerAccount);
+    repository.update(customerAccount);
 
     return Account.builder().balance(newBalance).build();
   }
 
   public Account withdraw(String id, double amount) {
-    Customer customerAccount = customerRepository.findById(id);
+    Customer customerAccount = repository.findById(id);
     if (customerAccount.getBalance() < amount) {
       throw new BadRequestException("You have insufficient founds");
     }
 
     double newBalance = customerAccount.getBalance() - amount;
     customerAccount.setBalance(newBalance);
-    customerRepository.update(customerAccount);
+    repository.update(customerAccount);
 
     return Account.builder().balance(newBalance).build();
   }
