+++ src/main/java/com/rflpazini/rogue/domain/usecase/CustomerService.java
@@ -1,13 +1,13 @@
 package com.rflpazini.rogue.domain.usecase;
 
-import com.rflpazini.rogue.app.dataprovider.impl.CustomerRepository;
-import com.rflpazini.rogue.app.dataprovider.model.Customer;
+import com.rflpazini.rogue.app.entrypoint.model.Customer;
+import com.rflpazini.rogue.domain.dataprovider.CrudPattern;
 import java.util.List;
 import javax.inject.Inject;
 
 public class CustomerService {
 
-  @Inject private CustomerRepository repository;
+  @Inject private CrudPattern<Customer> repository;
 
   public List<Customer> getAll() {
     return repository.findAll();
