--- src/main/resources/META-INF/microprofile-config.properties
@@ -1,2 +1,0 @@
-injected.value=Injected value
-value=lookup value
